requirements.txt
ipython==8.12.3
matplotlib==3.9.4
networkx==3.0
numpy==1.26.4
optuna==4.2.0
pandas==2.0.3
rdkit==2024.3.5
scikit_learn==1.3.2
scipy==1.10.1
torch==2.5.1+cu118
torch_geometric==2.6.1
umap_learn==0.5.7
pydot==4.0.1
seaborn==0.13.2